pyproject.toml
[project]
name = "depers-sys"
version = "0.1.0"
description = "Local AI-based data depersonalization system"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "python-multipart>=0.0.9",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "transformers>=4.40",
    "torch>=2.2",
    "Pillow>=10.2",
    "pandas>=2.2",
    "openpyxl>=3.1",
    "python-docx>=1.1",
    "PyPDF2>=3.0",
    "opencv-python-headless>=4.9",
    "easyocr>=1.7",
    "aiofiles>=23.2",
    "loguru>=0.7",
]

[project.optional-dependencies]
dev = ["pytest>=8.1", "pytest-asyncio>=0.23", "httpx>=0.27", "ruff>=0.4"]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["frontend*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
